Renamed folder to collection-form
